+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+### Added
+
+- Add support for listing address transactions.
+
+## [0.2.0]
+
+### Added
+
+- USDC Faucet support on Base Sepolia
+- Improved error mesasges for `InternalError`
+
+## [0.1.1] - 2024-08-27
+
+- Fixed a bug where `listHistoricalBalances` method was parsing conventional ETH balances instead of atomic units
+
 ## [0.1.0] - 2024-08-22
 
 ### Added